+++ pkgs/native_assets_builder/test_data/complex_link_helper/hook/build.dart
@@ -2,33 +2,37 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:io';
+
 import 'package:native_assets_cli/native_assets_cli.dart';
 
-const packageName = 'complex_link_helper';
+void main(List<String> args) async {
+  build(args, (config, output) async {
+    final packageName = config.packageName;
+    final assetDirectory =
+        Directory.fromUri(config.packageRoot.resolve('assets/'));
+    // If assets are added, rerun hook.
+    output.addDependency(assetDirectory.uri);
 
-void main(List<String> args) async => build(args, (config, output) async {
-      output.addAssets(
-        List.generate(
-          2,
-          (index) => DataAsset(
-            name: 'data_helper_$index',
-            // TODO(mosuem): Simplify specifying files/file paths
-            file: config.packageRoot.resolve('assets/data_helper_$index.json'),
-            package: packageName,
-          ),
+    await for (final dataAsset in assetDirectory.list()) {
+      if (dataAsset is! File) {
+        continue;
+      }
+      final fileName = dataAsset.uri.pathSegments.last;
+      final name = fileName.split('.').first;
+      final forLinking = name.contains('2') || name.contains('3');
+      output.addAsset(
+        DataAsset(
+          package: packageName,
+          name: name,
+          file: dataAsset.uri,
         ),
+        linkInPackage:
+            forLinking && config.linkingEnabled ? 'complex_link' : null,
       );
-      output.addAssets(
-        List.generate(
-          2,
-          (index) => DataAsset(
-            name: 'data_helper_${index + 2}',
-            // TODO(mosuem): Simplify specifying files/file paths
-            file: config.packageRoot
-                .resolve('assets/data_helper_${index + 2}.json'),
-            package: packageName,
-          ),
-        ),
-        linkInPackage: config.linkingEnabled ? 'complex_link' : null,
-      );
-    });
+      // TODO(https://github.com/dart-lang/native/issues/1208): Report
+      // dependency on asset.
+      output.addDependency(dataAsset.uri);
+    }
+  });
+}