+++ pkgs/native_assets_builder/test_data/simple_data_asset/hook/build.dart
@@ -1,18 +1,35 @@
 // Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:io';
+
 import 'package:native_assets_cli/native_assets_cli.dart';
 
-void main(List<String> args) => build(
-      args,
-      (config, output) async {
+void main(List<String> args) async {
+  await build(
+    args,
+    (config, output) async {
+      final assetDirectory =
+          Directory.fromUri(config.packageRoot.resolve('assets/'));
+      // If assets are added, rerun hook.
+      output.addDependency(assetDirectory.uri);
+
+      await for (final dataAsset in assetDirectory.list()) {
+        if (dataAsset is! File) {
+          continue;
+        }
+        final name = dataAsset.uri.pathSegments.last;
         output.addAsset(
           DataAsset(
             package: config.packageName,
-            name: 'assetId',
-            file: config.packageRoot.resolve('asset/test_asset.txt'),
+            name: name,
+            file: dataAsset.uri,
           ),
         );
-        output.addDependency(config.packageRoot.resolve('hook/build.dart'));
-      },
-    );
+        // TODO(https://github.com/dart-lang/native/issues/1208): Report
+        // dependency on asset.
+        output.addDependency(dataAsset.uri);
+      }
+    },
+  );
+}