+++ pkgs/native_assets_cli/example/link/package_with_assets/hook/build.dart
@@ -2,26 +2,35 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:io';
+
 import 'package:native_assets_cli/native_assets_cli.dart';
 
 void main(List<String> args) async {
   await build(args, (config, output) async {
     final packageName = config.packageName;
-    final allAssets = [
-      DataAsset(
-        package: packageName,
-        name: 'unused_asset',
-        file: config.packageRoot.resolve('assets/unused_asset.json'),
-      ),
-      DataAsset(
-        package: packageName,
-        name: 'used_asset',
-        file: config.packageRoot.resolve('assets/used_asset.json'),
-      )
-    ];
-    output.addAssets(
-      allAssets,
-      linkInPackage: config.linkingEnabled ? packageName : null,
-    );
+    final assetDirectory =
+        Directory.fromUri(config.packageRoot.resolve('assets/'));
+    // If assets are added, rerun hook.
+    output.addDependency(assetDirectory.uri);
+
+    await for (final dataAsset in assetDirectory.list()) {
+      if (dataAsset is! File) {
+        continue;
+      }
+      final fileName = dataAsset.uri.pathSegments.last;
+      final name = fileName.split('.').first;
+      output.addAsset(
+        DataAsset(
+          package: packageName,
+          name: name,
+          file: dataAsset.uri,
+        ),
+        linkInPackage: config.linkingEnabled ? packageName : null,
+      );
+      // TODO(https://github.com/dart-lang/native/issues/1208): Report
+      // dependency on asset.
+      output.addDependency(dataAsset.uri);
+    }
   });
 }