+++ pkgs/native_assets_cli/test/example/local_asset_test.dart
@@ -83,7 +83,6 @@ void main() async {
           dependencies,
           [
             testPackageUri.resolve('data/asset.txt'),
-            testPackageUri.resolve('hook/build.dart'),
           ],
         );
       }